* as/asranlib/asranlib.h: changed type of offset to unsigned int to   make it work on ppc Mac OS X (off_t is 64 bit and is not converted   to 32 bit unsigned int when calling sputl() since it is a macro * as/link/lkar.c: changed the initialization of This->loaded
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@5348 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ as/asranlib/asranlib.c
@@ -93,7 +93,7 @@ struct symbol_s
   };
 
 struct symbol_s *symlist, *lastsym;
-off_t offset, first_member_offset;
+unsigned int offset, first_member_offset;
 
 int
 add_symbol (const char *sym, void *param)
